backend/requirements.txt
a2wsgi==1.10.10
ads==0.12.7
aerospike_vector_search==4.2.0
ain==0.0.2
aiocontextvars==0.2.2
aiodns==3.5.0
airbyte_cdk==6.58.0
aleph_alpha_client==10.4.0
alibabacloud_ha3engine_vector==1.1.17
alibabacloud_tea_util==0.3.13
apify_client==1.12.0
arango==0.2.1
arcgis==2.4.1.1
argilla==2.8.0
arize==7.47.0
arthurai==3.33.3
astrapy==2.0.1
async_timeout==5.0.1
atlassian==0.0.0
azureml==0.2.7
baidubce==0.8.3
banana_dev==6.3.0
bibtexparser==1.4.3
bigdl==2.4.0
bilibili_api==9.1.0
blobfile==3.0.0
boto3==1.39.3
botocore==1.39.3
brotli==1.1.0
brotlicffi==1.1.0.0
cassio==0.1.10
chardet==5.2.0
checks==0.2
chromadb==1.0.15
clarifai_grpc==11.6.1
clickhouse_connect==0.8.18
cloudpathlib==0.21.1
cloudpickle==3.1.1
colorama==0.4.6
comet_llm==2.2.7
confluent_kafka==2.11.0
cryptography==45.0.5
curio==1.6
cx_Oracle==8.3.0
Cython==3.1.2
datadog_api_client==2.39.0
dbruntime==0.0.1
deepeval==3.2.4
defusedxml==0.7.1
dgml_utils==0.3.4
dingodb==0.0.18
dl==0.1.0
doctran==0.0.14
docutils==0.21.2
docx2txt==0.9
dotenv
dria==0.0.124
dynamodb_encryption_sdk==3.3.0
e2b==1.5.4
elevenlabs==2.6.0
email_validator==2.2.0
esprima==4.0.1
eval_type_backport==0.2.2
exceptiongroup==1.3.0
extract_msg==0.54.1
fake_useragent==2.2.0
faiss-cpu
fastapi
fastapi_cli==0.0.8
faster_whisper==1.1.1
feedparser==6.0.11
filelock==3.18.0
firebase_admin==6.9.0
fitz==0.0.1.dev2
flashrank==0.2.10
flytekit==1.16.1
geopandas==1.1.1
gliner==0.2.21
google_api_python_client==2.176.0
google_auth_oauthlib==1.2.2
googlemaps==4.10.0
gptcache==0.1.44
gpudb==7.2.2.9
gql==3.5.3
gradientai==1.13.1
grandalf==0.8
graphviz==0.21
grpcio==1.57.0
guardrails==2.0.0
gunicorn==23.0.0
h2==4.2.0
h5py==3.14.0
hdbcli==2.25.22
hologres_vector==0.0.10
httptools==0.6.4
httpx_aiohttp==0.1.8
hypothesis==6.135.26
ibm_watsonx_ai==1.3.30
importlib_resources==6.5.2
infinity_client==0.0.76
infinity_emb==0.0.76
ini2toml==0.15
InstructorEmbedding==1.0.1
intel_extension_for_transformers
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.7
itsdangerous==2.2.0
jaguardb_http_client==3.4.5
javelin_sdk==0.2.37
Jinja2==3.1.6
jnius==1.1.0
joblib==1.5.1
jq==1.9.1
js==1.0
jsonschema==4.24.0
kdbai_client==1.7.0
keybert==0.9.0
label_studio_sdk==1.0.18
langchain
langchain_openai
distro>=1.8.0
langchain_anthropic==0.3.17
langchain_astradb==0.6.0
langchain_aws==0.2.27
langchain_azure_ai==0.1.4
langchain_chroma==0.2.4
langchain_cohere==0.4.4
langchain_community==0.3.27
langchain_deepseek==0.1.3
langchain_elasticsearch==0.3.2
langchain_experimental==0.3.4
langchain_fireworks==0.3.0
langchain_google_genai==2.1.6
langchain_google_vertexai==2.0.27
langchain_groq==0.3.5
langchain_huggingface==0.3.0
langchain_ibm==0.3.15
langchain_milvus==0.2.1
langchain_mistralai==0.2.11
langchain_mongodb==0.6.2
langchain_neo4j==0.4.0
langchain_nvidia_ai_endpoints==0.3.10
langchain_ollama==0.3.4
langchain_perplexity==0.1.1
langchain_pinecone==0.2.8
langchain_postgres==0.0.15
langchain_qdrant==0.2.0
langchain_together==0.3.0
langchain_weaviate==0.0.5
langchain_xai==0.2.4
langchainhub==0.1.21
langsmith_pyo3==0.1.0rc5
lark==1.2.2
laser_encoders==0.0.2
librosa==0.11.0
listparser==0.20
llmlingua==0.2.2
llmonitor==0.0.51
lorax==0.94.86
lxml==6.0.0
mangum
manticoresearch==8.0.0
mask==1.0.0
matplotlib==3.10.3
mediawikiapi==1.3
metal_sdk==2.5.1
mlx_lm==0.26.0
modelscope==1.28.0
modin==0.34.0
motor==3.7.1
msal==1.32.3
mtrand==0.1
mwparserfromhell==0.7.2
mwxml==0.3.6
mypy_boto3_secretsmanager==1.39.0
nest_asyncio==1.6.0
nlu==5.4.1
nomic==3.5.3
nucliadb_models==6.6.1.post4652
nucliadb_protos==6.6.1.post4652
numexpr==2.11.0
O365==2.1.4
oci==2.155.1
odps==3.5.1
openapi_pydantic==0.5.1
optimum==1.26.1
outcome==1.3.0.post0
pdf2image==1.17.0
pdfminer==20191125
pdfplumber==0.11.7
Pillow==11.3.0
pinecone_text==0.10.0
pkg1==0.0.3
polars==1.31.0
praw==7.8.1
promptlayer==1.0.60
psutil==7.0.0
psychicapi==0.8.4
pyairtable==3.1.1
pycryptodome==3.23.0
pydub==0.25.1
pyepsilla==0.3.14
Pygments==2.19.2
pygraphviz==1.14
PyInstaller==6.14.2
PyJWT==2.10.1
pymemcache==4.0.0
pymilvus==2.5.12
pymongo==4.13.2
pymupdf==1.26.3
pyOpenSSL==25.1.0
pyowm==3.3.0
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
pyppeteer==2.0.0
pyspark==4.0.0
pysqlcipher3==1.2.0
pysrt==1.1.2
pytesseract==0.3.13
pytest==8.4.1
python-decouple==3.8
python_magic==0.4.27
python_multipart==0.0.20
pytidylib==0.3.2
pyTigerGraph==1.9.0
pytube==15.0.0
pytz==2025.2
qdrant_client==1.14.3
qianfan==0.4.12.3
rank_bm25==0.2.2
rank_llm==0.21.0
rapidfuzz==3.13.0
rapidocr_onnxruntime==1.4.4
rdflib==7.1.4
reportlab==4.4.2
rizaio==0.12.0
rockset==2.1.2
rspace_client==2.6.1
runhouse==0.0.42
scipy==1.16.0
scipy_doctest==1.8.0
scrapingant_client==2.1.0
selenium==4.34.2
semantic_version==2.10.0
setuptools_scm==8.3.1
simplejson==3.20.1
simsimd==6.5.0
slack_sdk==3.35.0
socksio==1.0.0
sodapy==2.2.0
sounddevice==0.5.2
speechkit==2.2.3
sqlcipher3==0.5.4
sqlite_vec==0.1.6
sqlite_vss==0.1.2
sseclient==0.0.27
stackapi==0.3.1
steamship==2.17.34
steamspypi==1.1.1
svglib==1.5.1
tensorflow
tensorflow_text
text_generation==0.7.0
textractor==0.1.2
thirdai==0.9.33
threadpoolctl==3.6.0
tigrisdb==1.0.0b6
timescale_vector==0.0.7
tokenizers==0.21.2
tomli_w==1.2.0
torch_npu==2.5.1
tornado==6.5.1
tree_sitter==0.24.0
tree_sitter_languages==1.10.2
trove_classifiers==2025.5.9.12
trubrics==1.8.5
tweepy==4.16.0
ujson==5.10.0
upstash_ratelimit==1.1.0
upstash_vector==0.8.0
vearch_cluster==0.0.2
watchfiles==1.1.0
websocket_client==1.8.0
whylogs==1.6.4
wikibase_rest_api_client==0.2.5
win32_setctime==1.2.0
wolframalpha==5.1.3
writerai==4.0.1
wsproto==1.2.0
xinference_client==1.7.1.post1
xlrd==2.0.2
xmltodict==0.14.2
xx==3.3.2
yfinance==0.2.65
youtube_search==2.1.2
youtube_transcript_api==1.1.1
yt_dlp==2025.6.30
zep_python==2.0.2
